refactor: move constants to libs
