+++ tests/models/test_extra.py
@@ -3,13 +3,13 @@ from unittest.mock import patch
 
 from django.test import SimpleTestCase
 
-from app.constants import Direction, RankName
 from app.models.basho import Basho
 from app.models.bout import Bout
 from app.models.division import Division
 from app.models.history import BashoHistory
 from app.models.rank import Rank
 from app.models.rikishi import Heya, Rikishi, Shusshin
+from libs.constants import Direction, RankName
 
 
 class ModelUtilityTests(SimpleTestCase):
