+++ tests/models/test_rikishi_ordering.py
@@ -1,10 +1,10 @@
 from django.test import TestCase
 from django.urls import reverse
 
-from app.constants import Direction, RankName
 from app.models.division import Division
 from app.models.rank import Rank
 from app.models.rikishi import Rikishi
+from libs.constants import Direction, RankName
 
 
 class RikishiOrderingTests(TestCase):
