+++ tests/views/test_rikishi_list.py
@@ -3,10 +3,10 @@ from datetime import date
 from django.test import TestCase
 from django.urls import reverse
 
-from app.constants import Direction, RankName
 from app.models.division import Division
 from app.models.rank import Rank
 from app.models.rikishi import Heya, Rikishi, Shusshin
+from libs.constants import Direction, RankName
 
 
 class RikishiListViewTests(TestCase):
